# docs/contracts/memongo/2.0.1/openapi.json
{
	"components": {
		"schemas": {
			"ApiError": {
				"properties": {
					"error": {
						"properties": {
							"code": {
								"type": "string"
							},
							"message": {
								"type": "string"
							}
						},
						"required": ["code", "message"],
						"type": "object"
					}
				},
				"required": ["error"],
				"type": "object"
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"scheme": "bearer",
				"type": "http"
			}
		}
	},
	"info": {
		"description": "HTTP API for the Memongo memory platform. Configure it with MEMONGO_MONGODB_URI and, optionally, ~/.memongo/memongo.json.",
		"title": "Memongo API",
		"version": "2.0.1"
	},
	"openapi": "3.0.3",
	"paths": {
...
		"/v1/add": {
			"post": {
				"parameters": [
					{
						"description": "Idempotency key for this write (IETF/Stripe semantics): a replay returns the original receipt, a key reused with a different payload yields 422 IDEMPOTENCY_CONFLICT. The Idempotency-Key header takes precedence when both are sent.",
						"in": "header",
						"name": "Idempotency-Key",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"agentId": {
										"type": "string"
									},
									"containerTag": {
										"deprecated": true,
										"description": "Deprecated compatibility alias for sessionId.",
										"type": "string"
									},
									"content": {
										"type": "string"
									},
									"customId": {
										"description": "Idempotency key for this write (IETF/Stripe semantics): a replay returns the original receipt, a key reused with a different payload yields 422 IDEMPOTENCY_CONFLICT. The Idempotency-Key header takes precedence when both are sent.",
										"type": "string"
									},
									"expiresAt": {
										"description": "ISO 8601 instant after which this document expires (P4.4.1 TTL). Must be in the future; an invalid or past value is rejected with 400 VALIDATION_ERROR.",
										"format": "date-time",
										"type": "string"
									},
									"metadata": {
										"description": "Optional metadata stored verbatim with the event (query-operator keys are rejected).",
										"type": "object"
									},
									"scope": {
										"description": "Optional memory isolation scope for retrieval.",
										"enum": [
											"session",
											"user",
											"agent",
											"workspace",
											"tenant",
											"global"
										],
										"type": "string"
									},
									"scopeRef": {
										"description": "Optional scope reference, for example a workspace path.",
										"type": "string"
									},
									"sessionId": {
										"description": "Optional session identifier for this memory write.",
										"type": "string"
									}
								},
								"required": ["content"],
								"type": "object"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Event id"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						},
						"description": "Validation error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						},
						"description": "Internal server error"
					}
				},
				"summary": "Append a user message to conversational memory"
			}
		},
...
		"/v1/context-bundle": {
			"post": {
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"agentId": {
										"type": "string"
									},
									"discoveryKind": {
										"enum": [
											"entity-brief",
											"topic-brief",
											"what-changed",
											"contradiction-report"
										],
										"type": "string"
									},
									"includeDiscoveryProjection": {
										"type": "boolean"
									},
									"includeProfile": {
										"type": "boolean"
									},
									"maxActiveItems": {
										"type": "number"
									},
									"maxEvidenceItems": {
										"type": "number"
									},
									"maxRecentEvents": {
										"type": "number"
									},
									"mode": {
										"description": "wake-up returns a compact session-start projection and skips query evidence",
										"enum": ["full", "wake-up"],
										"type": "string"
									},
									"query": {
										"type": "string"
									},
									"scope": {
										"enum": [
											"session",
											"user",
											"agent",
											"workspace",
											"tenant",
											"global"
										],
										"type": "string"
									},
									"scopeRef": {
										"type": "string"
									},
									"sessionId": {
										"type": "string"
									},
									"timeRange": {
										"properties": {
											"end": {
												"format": "date-time",
												"type": "string"
											},
											"preset": {
												"type": "string"
											},
											"start": {
												"format": "date-time",
												"type": "string"
											}
										},
										"type": "object"
									},
									"tokenBudget": {
										"type": "number"
									}
								},
								"type": "object"
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"agentId": {
											"type": "string"
										},
										"builtAt": {
											"format": "date-time",
											"type": "string"
										},
										"metadata": {
											"properties": {
												"estimatedTokensUsed": {
													"type": "number"
												},
												"partial": {
													"type": "boolean"
												},
												"pathsExecuted": {
													"items": {
														"type": "string"
													},
													"type": "array"
												},
												"sectionsIncluded": {
													"items": {
														"type": "string"
													},
													"type": "array"
												},
												"tokenBudget": {
													"type": "number"
												},
												"truncated": {
													"type": "boolean"
												}
											},
											"type": "object"
										},
										"query": {
											"type": "string"
										},
										"rendered": {
											"type": "string"
										},
										"scope": {
											"type": "string"
										},
										"scopeRef": {
											"type": "string"
										},
										"sections": {
											"items": {
												"properties": {
													"estimatedTokens": {
														"type": "number"
													},
													"items": {
														"items": {
															"properties": {
																"canonicalId": {
																	"type": "string"
																},
																"path": {
																	"type": "string"
																},
																"scope": {
																	"type": "string"
																},
																"scopeRef": {
																	"type": "string"
																},
																"source": {
																	"type": "string"
																},
																"sourceEventIds": {
																	"items": {
																		"type": "string"
																	},
																	"type": "array"
																},
																"summary": {
																	"type": "string"
																},
																"timestamp": {
																	"format": "date-time",
																	"type": "string"
																},
																"title": {
																	"type": "string"
																}
															},
															"type": "object"
														},
														"type": "array"
													},
													"kind": {
														"type": "string"
													},
													"partial": {
														"type": "boolean"
													},
													"summary": {
														"type": "string"
													},
													"title": {
														"type": "string"
													},
													"truncated": {
														"type": "boolean"
													}
												},
												"type": "object"
											},
											"type": "array"
										},
										"sessionId": {
											"type": "string"
										}
									},
									"type": "object"
								}
							}
						},
						"description": "Prompt-ready context bundle with structured sections"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						},
						"description": "Validation error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						},
						"description": "Internal server error"
					}
				},
				"summary": "Build a prompt-ready context bundle from active memory, durable evidence, summaries, and recent events"
			}
		},
		"/v1/discovery-projection": {
			"post": {
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"agentId": {
										"type": "string"
									},
									"kind": {
										"enum": [
											"entity-brief",
											"topic-brief",
											"what-changed",
											"contradiction-report"
										],
										"type": "string"
									},
									"maxItems": {
										"type": "number"
									},
									"query": {
										"type": "string"
									},
									"scope": {
										"enum": [
											"session",
											"user",
											"agent",
											"workspace",
											"tenant",
											"global"
										],
										"type": "string"
									},
									"scopeRef": {
										"type": "string"
									},
									"timeRange": {
										"properties": {
											"end": {
												"format": "date-time",
												"type": "string"
											},
											"preset": {
												"type": "string"
											},
											"start": {
												"format": "date-time",
												"type": "string"
											}
										},
										"type": "object"
									}
								},
								"required": ["kind"],
								"type": "object"
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"builtAt": {
											"format": "date-time",
											"type": "string"
										},
										"kind": {
											"type": "string"
										},
										"metadata": {
											"properties": {
												"evidenceCount": {
													"type": "number"
												},
												"partial": {
													"type": "boolean"
												},
												"sourceCounts": {
													"additionalProperties": {
														"type": "number"
													},
													"type": "object"
												},
												"timeRange": {
													"properties": {
														"end": {
															"format": "date-time",
															"type": "string"
														},
														"label": {
															"type": "string"
														},
														"start": {
															"format": "date-time",
															"type": "string"
														}
													},
													"type": "object"
												}
											},
											"type": "object"
										},
										"query": {
											"type": "string"
										},
										"scope": {
											"type": "string"
										},
										"scopeRef": {
											"type": "string"
										},
										"sections": {
											"items": {
												"properties": {
													"evidence": {
														"items": {
															"properties": {
																"canonicalId": {
																	"type": "string"
																},
																"path": {
																	"type": "string"
																},
																"source": {
																	"type": "string"
																},
																"summary": {
																	"type": "string"
																},
																"timestamp": {
																	"format": "date-time",
																	"type": "string"
																},
																"title": {
																	"type": "string"
																}
															},
															"type": "object"
														},
														"type": "array"
													},
													"summary": {
														"type": "string"
													},
													"title": {
														"type": "string"
													}
												},
												"type": "object"
											},
											"type": "array"
										},
										"summary": {
											"type": "string"
										},
										"title": {
											"type": "string"
										}
									},
									"type": "object"
								}
							}
						},
						"description": "Discovery projection with provenance-backed sections"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						},
						"description": "Validation error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						},
						"description": "Internal server error"
					}
				},
				"summary": "Build a rebuildable discovery projection such as an entity brief, topic brief, what-changed brief, or contradiction report"
			}
		},
		"/v1/extract": {
			"post": {
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"agentId": {
										"type": "string"
									},
									"eventId": {
										"minLength": 1,
										"type": "string"
									},
									"scope": {
										"description": "Optional memory isolation scope for retrieval.",
										"enum": [
											"session",
											"user",
											"agent",
											"workspace",
											"tenant",
											"global"
										],
										"type": "string"
									},
									"scopeRef": {
										"description": "Optional scope reference, for example a workspace path.",
										"type": "string"
									}
								},
								"required": ["eventId"],
								"type": "object"
							}
						}
					},
					"required": true
				},
				"responses": {
					"202": {
						"description": "Extraction scheduled"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						},
						"description": "Validation error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						},
						"description": "Internal server error"
					}
				},
				"summary": "Schedule background extraction for one canonical event"
			}
		},
		"/v1/hydrate-active-slate": {
			"post": {
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"agentId": {
										"type": "string"
									},
									"maxItems": {
										"description": "Requested slate size. Clamped to 6 items.",
										"type": "number"
									},
									"scope": {
										"enum": [
											"session",
											"user",
											"agent",
											"workspace",
											"tenant",
											"global"
										],
										"type": "string"
									},
									"scopeRef": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"agentId": {
											"type": "string"
										},
										"hydratedAt": {
											"format": "date-time",
											"type": "string"
										},
										"items": {
											"items": {
												"properties": {
													"canonicalId": {
														"type": "string"
													},
													"kind": {
														"type": "string"
													},
													"path": {
														"type": "string"
													},
													"scope": {
														"type": "string"
													},
													"scopeRef": {
														"type": "string"
													},
													"source": {
														"type": "string"
													},
													"summary": {
														"type": "string"
													},
													"timestamp": {
														"format": "date-time",
														"type": "string"
													},
													"title": {
														"type": "string"
													}
												},
												"type": "object"
											},
											"type": "array"
										},
										"metadata": {
											"properties": {
												"countsByKind": {
													"additionalProperties": {
														"type": "number"
													},
													"type": "object"
												},
												"maxItems": {
													"type": "number"
												},
												"partial": {
													"type": "boolean"
												},
												"sourceCounts": {
													"additionalProperties": {
														"type": "number"
													},
													"type": "object"
												},
												"truncated": {
													"type": "boolean"
												}
											},
											"type": "object"
										},
										"scope": {
											"type": "string"
										},
										"scopeRef": {
											"type": "string"
										}
									},
									"type": "object"
								}
							}
						},
						"description": "Tiny active-memory slate"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						},
						"description": "Validation error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						},
						"description": "Internal server error"
					}
				},
				"summary": "Hydrate a tiny active-memory slate for recall-heavy turns and debugging surfaces"
			}
		},
...
		"/v1/profile": {
			"post": {
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"activityWindowMs": {
										"type": "number"
									},
									"agentId": {
										"type": "string"
									},
									"containerTag": {
										"deprecated": true,
										"description": "Deprecated compatibility alias for scopeRef.",
										"type": "string"
									},
									"maxEntities": {
										"type": "number"
									},
									"maxEpisodes": {
										"type": "number"
									},
									"maxPerType": {
										"type": "number"
									},
									"scope": {
										"description": "Optional scope for profile synthesis. Defaults to agent.",
										"enum": [
											"session",
											"user",
											"agent",
											"workspace",
											"tenant",
											"global"
										],
										"type": "string"
									},
									"scopeRef": {
										"description": "Optional scope reference for profile synthesis.",
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Profile"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						},
						"description": "Validation error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						},
						"description": "Internal server error"
					}
				},
				"summary": "Synthesize a profile for a scope"
			}
		},
...
	},
	"security": [
		{
			"bearerAuth": []
		}
	],
	"servers": [
		{
			"description": "Default",
			"url": "/"
		}
	]
}